+++ datacube/drivers/postgis/_api.py
@@ -24,7 +24,6 @@ from sqlalchemy.sql.expression import Select
 from sqlalchemy import select, text, and_, or_, func, column
 from sqlalchemy.dialects.postgresql import INTERVAL
 from sqlalchemy.exc import IntegrityError
-from sqlalchemy.engine import Row
 
 from typing import Iterable, Sequence, Optional, Set, Any
 from typing import cast as type_cast
@@ -50,22 +49,40 @@ _LOG = logging.getLogger(__name__)
 
 # Make a function because it's broken
 def _dataset_select_fields() -> tuple:
+    return tuple(f.alchemy_expression for f in _dataset_fields())
+
+
+def _dataset_fields() -> tuple:
+    native_flds = get_native_fields()
     return (
-        Dataset,
-        # All active URIs, from newest to oldest
-        func.array(
-            select(
-                SelectedDatasetLocation.uri
-            ).where(
-                and_(
-                    SelectedDatasetLocation.dataset_ref == Dataset.id,
-                    SelectedDatasetLocation.archived == None
-                )
-            ).order_by(
-                SelectedDatasetLocation.added.desc(),
-                SelectedDatasetLocation.id.desc()
-            ).label('uris')
-        ).label('uris')
+        native_flds["id"],
+        native_flds["indexed_time"],
+        native_flds["indexed_by"],
+        native_flds["product_id"],
+        native_flds["metadata_type_id"],
+        native_flds["metadata_doc"],
+        NativeField(
+            'archived',
+            'Archived date',
+            Dataset.archived
+        ),
+        NativeField("uris",
+                    "all uris",
+                    func.array(
+                        select(
+                            SelectedDatasetLocation.uri
+                        ).where(
+                            and_(
+                                SelectedDatasetLocation.dataset_ref == Dataset.id,
+                                SelectedDatasetLocation.archived == None
+                            )
+                        ).order_by(
+                            SelectedDatasetLocation.added.desc(),
+                            SelectedDatasetLocation.id.desc()
+                        ).label('uris')
+                    ),
+                    alchemy_table=Dataset.__table__  # type: ignore[attr-defined]
+        )
     )
 
 
@@ -228,6 +245,29 @@ def extract_dataset_fields(ds_metadata, fields):
         except UnindexableValue:
             continue
     return result
+
+
+# Min/Max aggregating time fields for temporal_extent methods
+time_min = DateDocField('acquisition_time_min',
+                        'Min of time when dataset was acquired',
+                        Dataset.metadata_doc,
+                        False,  # is it indexed
+                        offset=[
+                            ['properties', 'dtr:start_datetime'],
+                            ['properties', 'datetime']
+                        ],
+                        selection='least')
+
+
+time_max = DateDocField('acquisition_time_max',
+                        'Max of time when dataset was acquired',
+                        Dataset.metadata_doc,
+                        False,  # is it indexed
+                        offset=[
+                            ['properties', 'dtr:end_datetime'],
+                            ['properties', 'datetime']
+                        ],
+                        selection='greatest')
 
 
 class PostgisDbAPI:
@@ -476,19 +516,6 @@ class PostgisDbAPI:
             )
         return self._connection.execute(query).fetchall()
 
-    # Not currently implemented.
-    # def insert_dataset_source(self, classifier, dataset_id, source_dataset_id):
-        # r = self._connection.execute(
-        #     insert(DatasetSource).on_conflict_do_nothing(
-        #         index_elements=['classifier', 'dataset_ref']
-        #     ).values(
-        #         classifier=classifier,
-        #         dataset_ref=dataset_id,
-        #         source_dataset_ref=source_dataset_id
-        #     )
-        # )
-        # return r.rowcount > 0
-
     def archive_dataset(self, dataset_id):
         r = self._connection.execute(
             update(Dataset).where(
@@ -548,10 +575,10 @@ class PostgisDbAPI:
         ).fetchall()
 
     def get_derived_datasets(self, dataset_id):
-        raise NotImplementedError
+        raise NotImplementedError()
 
     def get_dataset_sources(self, dataset_id):
-        raise NotImplementedError
+        raise NotImplementedError()
 
     def search_datasets_by_metadata(self, metadata, archived):
         """
@@ -622,13 +649,8 @@ class PostgisDbAPI:
         assert source_exprs is None
         assert not with_source_ids
 
-        if select_fields:
-            select_columns = tuple(
-                f.alchemy_expression.label(f.name)
-                for f in select_fields
-            )
-        else:
-            select_columns = _dataset_select_fields()
+        if not select_fields:
+            select_fields = _dataset_fields()
 
         def _ob_exprs(o):
             if isinstance(o, str):
@@ -644,6 +666,10 @@ class PostgisDbAPI:
         else:
             order_by = []
 
+        select_columns = tuple(
+            f.alchemy_expression.label(f.name)
+            for f in select_fields
+        )
         if geom:
             SpatialIndex, spatialquery = self.geospatial_query(geom)
         else:
@@ -678,13 +704,22 @@ class PostgisDbAPI:
         :type with_source_ids: bool
         :type select_fields: tuple[datacube.drivers.postgis._fields.PgField]
         :type expressions: tuple[datacube.drivers.postgis._fields.PgExpression]
+
+        :return: An iterable of tuples of decoded values
         """
         assert source_exprs is None
         assert not with_source_ids
+        if select_fields is None:
+            select_fields = _dataset_fields()
         select_query = self.search_datasets_query(expressions, select_fields=select_fields, limit=limit,
                                                   geom=geom, archived=archived, order_by=order_by)
         _LOG.debug("search_datasets SQL: %s", str(select_query))
-        return self._connection.execute(select_query)
+
+        def decode_row(raw: Iterable[Any]) -> dict[str, Any]:
+            return {f.name: f.normalise_value(r) for r, f in zip(raw, select_fields)}
+
+        for row in self._connection.execute(select_query):
+            yield decode_row(row)
 
     def bulk_simple_dataset_search(self, products=None, batch_size=0):
         """
@@ -706,7 +741,7 @@ class PostgisDbAPI:
         query = select(
             *_dataset_bulk_select_fields()
         ).select_from(Dataset).where(
-            Dataset.archived == None
+            Dataset.archived.is_(None)
         )
         if products:
             query = query.where(Dataset.product_ref.in_(products))
@@ -749,10 +784,12 @@ class PostgisDbAPI:
         )
         return res.rowcount, requested - res.rowcount
 
-    def get_duplicates(self, match_fields: Sequence[PgField], expressions: Sequence[PgExpression]) -> Iterable[Row]:
+    def get_duplicates(self,
+                       match_fields: Sequence[PgField],
+                       expressions: Sequence[PgExpression]) -> Iterable[dict[str, Any]]:
         # TODO
         if "time" in [f.name for f in match_fields]:
-            return self.get_duplicates_with_time(match_fields, expressions)
+            yield from self.get_duplicates_with_time(match_fields, expressions)
 
         group_expressions = tuple(f.alchemy_expression for f in match_fields)
         join_tables = PostgisDbAPI._join_tables(expressions, match_fields)
@@ -765,27 +802,31 @@ class PostgisDbAPI:
             query = query.join(*joins)
 
         query = query.where(
-            and_(Dataset.archived == None, *(PostgisDbAPI._alchemify_expressions(expressions)))
+            and_(Dataset.archived.is_(None), *(PostgisDbAPI._alchemify_expressions(expressions)))
         ).group_by(
             *group_expressions
         ).having(
             func.count(Dataset.id) > 1
         )
-        return self._connection.execute(query)
+        for row in self._connection.execute(query):
+            drow = {"ids": row.ids}
+            for f in match_fields:
+                drow[f.name] = getattr(row, f.name)
+            yield drow
 
     def get_duplicates_with_time(
             self, match_fields: Sequence[PgField], expressions: Sequence[PgExpression]
-    ) -> Iterable[Row]:
+    ) -> Iterable[dict[str, Any]]:
         fields = []
-        for f in match_fields:
-            if f.name == "time":
-                time_field = type_cast(DateRangeDocField, f).expression_with_leniency
+        for fld in match_fields:
+            if fld.name == "time":
+                time_field = type_cast(DateRangeDocField, fld)
             else:
-                fields.append(f.alchemy_expression)
+                fields.append(fld.alchemy_expression)
 
         join_tables = PostgisDbAPI._join_tables(expressions, match_fields)
 
-        cols = [Dataset.id, time_field.label('time'), *fields]
+        cols = [Dataset.id, time_field.expression_with_leniency.label('time'), *fields]
         query = select(
             *cols
         ).select_from(Dataset)
@@ -793,7 +834,7 @@ class PostgisDbAPI:
             query = query.join(*joins)
 
         query = query.where(
-            and_(Dataset.archived == None, *(PostgisDbAPI._alchemify_expressions(expressions)))
+            and_(Dataset.archived.is_(None), *(PostgisDbAPI._alchemify_expressions(expressions)))
         )
 
         t1 = query.alias("t1")
@@ -801,7 +842,7 @@ class PostgisDbAPI:
 
         time_overlap = select(
             t1.c.id,
-            text("t1.time * t2.time as time_intersect"),
+            t1.c.time.intersection(t2.c.time).label('time_intersect'),
             *fields
         ).select_from(
             t1.join(
@@ -813,7 +854,7 @@ class PostgisDbAPI:
         query = select(
             func.array_agg(func.distinct(time_overlap.c.id)).label("ids"),
             *fields,  # type: ignore[arg-type]
-            text("(lower(time_intersect) at time zone 'UTC', upper(time_intersect) at time zone 'UTC') as time")
+            text("time_intersect as time")
         ).select_from(
             time_overlap  # type: ignore[arg-type]
         ).group_by(
@@ -821,7 +862,16 @@ class PostgisDbAPI:
         ).having(
             func.count(time_overlap.c.id) > 1
         )
-        return self._connection.execute(query)
+
+        for row in self._connection.execute(query):
+            # TODO: Use decode_rows above - would require creating a field class for the ids array.
+            drow: dict[str, Any] = {
+                "ids": row.ids,
+            }
+            for f in fields:
+                drow[f.key] = getattr(row, f.key)  # type: ignore[union-attr]
+            drow["time"] = time_field.normalise_value((row.time.lower, row.time.upper))
+            yield drow
 
     def count_datasets(self, expressions, archived: bool | None = False, geom: Geometry | None = None):
         """
@@ -1371,7 +1421,7 @@ class PostgisDbAPI:
                                   source_id=rel.source_dataset_ref,
                                   derived_id=rel.derived_dataset_ref)
 
-    def write_relations(self, relations: Iterable[LineageRelation], allow_updates: bool):
+    def write_relations(self, relations: Iterable[LineageRelation], allow_updates: bool) -> int:
         """
         Write a set of LineageRelation objects to the database.
 
@@ -1379,6 +1429,7 @@ class PostgisDbAPI:
         :param allow_updates: if False, only allow adding new relations, not updating old ones.
         :return: Count of database rows affected
         """
+        affected = 0
         if allow_updates:
             by_classifier: dict[str, Any] = {}
             for rel in relations:
@@ -1391,32 +1442,31 @@ class PostgisDbAPI:
                     by_classifier[rel.classifier].append(db_repr)
                 else:
                     by_classifier[rel.classifier] = [db_repr]
-                updates = 0
                 for classifier, values in by_classifier.items():
                     qry = insert(DatasetLineage).on_conflict_do_update(
                         index_elements=["derived_dataset_ref", "source_dataset_ref"],
                         set_={"classifier": classifier},
                         where=(DatasetLineage.classifier != classifier))
                     res = self._connection.execute(qry, values)
-                    updates += res.rowcount
-                return updates
+                    affected += res.rowcount
         else:
-            values = [
-                {
-                    "derived_dataset_ref": rel.derived_id,
-                    "source_dataset_ref": rel.source_id,
-                    "classifier": rel.classifier
-                }
-                for rel in relations
-            ]
-            qry = insert(DatasetLineage)
-            try:
-                res = self._connection.execute(
-                    qry, values
-                )
-                return res.rowcount
-            except IntegrityError:
-                return 0
+            for rel in relations:
+                values = [
+                    {
+                        "derived_dataset_ref": rel.derived_id,
+                        "source_dataset_ref": rel.source_id,
+                        "classifier": rel.classifier
+                    }
+                ]
+                qry = insert(DatasetLineage)
+                try:
+                    res = self._connection.execute(
+                        qry, values
+                    )
+                    affected += res.rowcount
+                except IntegrityError:
+                    return 0
+        return affected
 
     def load_lineage_relations(self,
                                roots: Iterable[uuid.UUID],
@@ -1490,33 +1540,20 @@ class PostgisDbAPI:
     def temporal_extent_by_prod(self, product_id: int) -> tuple[datetime.datetime, datetime.datetime]:
         query = self.temporal_extent_full().where(Dataset.product_ref == product_id)
         res = self._connection.execute(query)
-        return res.first()
+        for tmin, tmax in res:
+            return (time_min.normalise_value(tmin), time_max.normalise_value(tmax))
+        raise RuntimeError("Product has no datasets and therefore no temporal extent")
 
     def temporal_extent_by_ids(self, ids: Iterable[DSID]) -> tuple[datetime.datetime, datetime.datetime]:
         query = self.temporal_extent_full().where(Dataset.id.in_(ids))
         res = self._connection.execute(query)
-        return res.first()
+        for tmin, tmax in res:
+            return (time_min.normalise_value(tmin), time_max.normalise_value(tmax))
+        raise ValueError("no dataset ids provided")
 
     def temporal_extent_full(self) -> Select:
         # Hardcode eo3 standard time locations - do not use this approach in a legacy index driver.
-        time_min = DateDocField('aquisition_time_min',
-                                'Min of time when dataset was acquired',
-                                Dataset.metadata_doc,
-                                False,  # is it indexed
-                                offset=[
-                                    ['properties', 'dtr:start_datetime'],
-                                    ['properties', 'datetime']
-                                ],
-                                selection='least')
-        time_max = DateDocField('aquisition_time_max',
-                                'Max of time when dataset was acquired',
-                                Dataset.metadata_doc,
-                                False,  # is it indexed
-                                offset=[
-                                    ['properties', 'dtr:end_datetime'],
-                                    ['properties', 'datetime']
-                                ],
-                                selection='greatest')
+
         return select(
             func.min(time_min.alchemy_expression), func.max(time_max.alchemy_expression)
         )