+++ docs/about/whats_new.rst
@@ -14,6 +14,12 @@ v1.9.next
 - Remove workaround for an odc-geo bug that is now fixed. (:pull:`1622`)
 - Fix call to geopolygon search. (:pull:`1627`)
 - Use antimeridian package to "fix" extent polygons. (:pull:`1628`)
+- Record lineage when adding datasets with postgis index (:pull:`1632`)
+- Update schema logic (:pull:`1634`)
+- Drop valid-area check and anti-meridian fix 3857 extents (:pull:1635)
+- Remove problematic "common_timestamp" postgresql function from postgis driver. Some internal API changes
+  required to accommodate and preserve all previous index-driver level behaviour. (:pull:`1623`)
+- Cherry picks from 1.8 (#1624-#1626, #1629-#1631) (:pull:`1637`)
 
 v1.9.0-rc9 (3rd July 2024)
 ==========================
@@ -132,6 +138,7 @@ v1.9.0-rc1 (27th March 2024)
 
 v1.8.next
 =========
+- Don't error when adding a dataset whose product doesn't have an id value (:pull:`1630`)
 
 v1.8.19 (2nd July 2024)
 =======================